Changed the geo dashboard and break PagarBookGeo.jsx file into component
